backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.37.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.12.0
PyJWT==2.8.0
openai-whisper==20250625
google-generativeai==0.8.3
pytube==15.0.0
moviepy==2.2.1
requests==2.32.5
python-dotenv==1.1.1
slowapi==0.1.9
email-validator==2.1.0
boto3==1.35.36
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
yt-dlp==2025.09.01
numpy==1.26.4
redis==5.1.0
rq==1.22.0
faster-whisper==0.11.0
opencv-python-headless==4.7.0.72
pillow==10.1.0
python-socketio==5.9.0
